refactor: address architecture review candidates #1, #2, #3, #4, and #6
