+++ src/prompt-input.test.ts
@@ -1,0 +1,9 @@
+import { describe, expect, test } from "bun:test";
+import { promptForInput } from "./prompt-input";
+
+describe("promptForInput", () => {
+  test("is exported and callable", () => {
+    expect(typeof promptForInput).toBe("function");
+    expect(promptForInput.length).toBe(1);
+  });
+});
